Branch b1_8 b=17336 i=robert.read i=jinshan.xiong
Do not trigger readpage when objective page index exceeds the end-of-file page index.

+++ lustre/liblustre/rw.c
@@ -707,7 +707,7 @@ ssize_t llu_file_prwv(const struct iovec *iovec, int iovlen,
                         } else {
                                 /* If objective page index exceed end-of-file
                                  * page index, return directly. --bug 17336 */
-                                size_t size = st->st_size;
+                                loff_t size = st->st_size;
                                 unsigned long cur_index = pos >> CFS_PAGE_SHIFT;
 
                                 if ((size == 0 && cur_index != 0) ||
